Return to the edit user page if user cannot be updated.
