+++ test/controllers/UpdateUserTest.scala
@@ -1,9 +1,10 @@
 package controllers
 
-import models.{User, UserUpdateRequest}
+import models.{Forms, User, UserUpdateRequest}
 import org.mockito.Mockito._
 import org.scalatest.mock.MockitoSugar
 import org.scalatestplus.play.{OneServerPerSuite, PlaySpec}
+import play.api.data.Form
 import play.api.test.Helpers._
 import services.{CustomError, AdminApi}
 
@@ -13,30 +14,32 @@ import scala.concurrent.Future
 class UpdateUserTest extends PlaySpec with OneServerPerSuite with MockitoSugar{
 
   val adminApiMock = mock[AdminApi]
+
   val controller = new SaveAction {override val adminApi: AdminApi = adminApiMock}
+
   val searchQuery = "search query"
   val userId = "1234"
   val userUpdateData = UserUpdateRequest("email@email.com","username")
   val blankUser = User("","", username=Some(""))
+  val blankUserForm = Forms.createForm(blankUser)
   val customError = CustomError("Test Error", "There is an error")
 
-  "doSave" should {
+  "update" should {
 
     "redirect to search results with success message on success" in {
       when(adminApiMock.updateUserData(userId, userUpdateData)).thenReturn(
         Future.successful(Right(blankUser))
       )
-      val result = controller.doSave(userId, userUpdateData, searchQuery)
+      val result = controller.update(userId,searchQuery, userUpdateData, blankUserForm)
       redirectLocation(result) mustEqual Some(routes.Search.search(searchQuery).url)
       flash(result).get("message") mustEqual Some("User has been updated")
     }
 
-    "redirect to search results with error message on failure" in {
+    "when form input is invalid return to the edit user page" in {
       when(adminApiMock.updateUserData(userId, userUpdateData)).thenReturn(
         Future.successful(Left(customError)))
-      val result = controller.doSave(userId, userUpdateData, searchQuery)
-      redirectLocation(result) mustEqual Some(routes.Search.search(searchQuery).url)
-      flash(result).get("error") mustEqual Some("Test Error : There is an error")
+      val result = controller.update(userId, searchQuery, userUpdateData, blankUserForm)
+      status(result) mustEqual OK
     }
   }
 }
